Introduced a Spring Data Repository implementation and example (GemstoneRepository) in addition to and in replacement of the GemFireGemstoneDao implementation for the gemfire-mysql-example project module.

+++ gemfire-mysql-example/src/main/java/sample/app/service/provider/GemFireGemstoneService.java
@@ -92,7 +92,7 @@ public class GemFireGemstoneService extends GemstoneService {
   }
 
   @Override
-  public int countFromGemFire() {
+  public long countFromGemFire() {
     /*
     return doInTransaction(new TransactionCallback<Object, Integer>() {
       @Override public Integer doInTransaction(final Object object) {
